+++ domain/yolo/yolo_schema.py
